requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
mysql-connector-python
